requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
dj-rest-auth==7.0.0
Django==5.1.4
django-allauth==65.3.0
django-redis==5.4.0
django-rest-auth==0.9.5
django-stubs==5.1.1
django-stubs-ext==5.1.1
django-types==0.19.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
idna==3.10
importlib_resources==6.4.5
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
oauthlib==3.2.2
packaging==24.2
pycparser==2.22
PyJWT==2.10.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.22.3
six==1.17.0
sqlparse==0.5.2
swagger-spec-validator==3.0.4
types-psycopg2==2.9.21.20241019
types-PyYAML==6.0.12.20240917
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
